Create function build in app.js
This function is used to get the data from API then group them into students, lecturers, phdLecturers, profLecturers and HOD. students are users that are less than 55. lecturers are users that are greater than 54 and less than or equal to 60. phdLecturers are users that are greater than 60 and less than or equal to 65. profLecturers are users that are greater 65. HOD is the oldest profLecturers (user).