Creates appendElems function in app.js/.
This function is takes in two paramenters, role and arrObj which are role and array/objec. The function first determines if the arrObj is an array or an object. The aim is create HTML for each user to display in the year book. It calls another function *createElem* which should return a div using an obj and the role of arrObj.

If array, it loops through, extracting each item which is an object then call createElem with item object and the role of arrObj. But if just an Object, it calls createElem with arrObj and it's role.